refactor(web): extract SocialMedias component
