+++ web/components/MobileMenu/MobileMenu.module.scss
@@ -90,18 +90,6 @@
     }
   }
 
-  .socialMedias {
-    display: flex;
-    justify-content: center;
-    gap: 26px;
-    margin-top: 100px;
-    & > * {
-      &:hover {
-        opacity: 0.7;
-      }
-    }
-  }
-
   .join {
     all: unset;
     &:hover {
